MicroBlog/requirements.txt
aiohttp==3.3.2
aiomysql==0.0.19
alembic==1.0.0
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.0
attrs==18.1.0
Babel==2.6.0
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
cryptography==2.3
dominate==2.3.1
<<<<<<< HEAD
=======
elasticsearch==6.3.1
>>>>>>> c42c925d5050d8a96c495134a64f051b09ad0d43
et-xmlfile==1.0.1
excel==1.0.0
facebook-sdk==2.0.0
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-Stormpath==0.4.8
Flask-WTF==0.14.2
guess-language==0.2
guess-language-spirit==0.5.3
httplib2==0.11.3
idna==2.7
idna-ssl==1.1.0
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
multidict==4.3.1
mysql-connector-python==8.0.12
oauth2client==1.5.2
oauthlib==1.0.3
openpyxl==2.5.5
Pillow==5.2.0
protobuf==3.6.0
psutil==5.4.6
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyJWT==1.6.4
pylint==2.1.1
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
rsa==3.4.2
selenium==3.13.0
six==1.11.0
SQLAlchemy==1.2.10
stormpath==2.4.4
typed-ast==1.1.0
urllib3==1.23
virtualenv==16.0.0
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
yarl==1.2.6
